*fix bug: Compiler error in function Gemm32fNN (32 bit mode).

+++ src/Simd/SimdMotion.hpp
@@ -1,7 +1,7 @@
 /*
 * Simd Library (http://ermig1979.github.io/Simd).
 *
-* Copyright (c) 2011-2017 Yermalayeu Ihar.
+* Copyright (c) 2011-2018 Yermalayeu Ihar.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
@@ -1276,7 +1276,7 @@ namespace Simd
                 for (size_t j = 0; j < objects.size(); ++j)
                 {
                     ObjectPtr & object = objects[j];
-                    if (object->trajectory.size() > _options.TrackingTrajectoryMax)
+                    if (object->trajectory.size() > (size_t)_options.TrackingTrajectoryMax)
                         object->trajectory.erase(object->trajectory.begin());
                 }
             }
